% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMain.R, R/rideProfile.R
\name{rideProfile}
\alias{rideProfile}
\title{plot heartrate, cadence, speed and more using color-driven
  data summaries to convey performance patterns}
\usage{
rideProfile(
  track,
  summary,
  savefn,
  title = "Ride starting ",
  palette = "plasma",
  naPlotColor = "gray88",
  verticalMultiplier = NA,
  ppm = NA,
  displayBands = c("speed", "grade", "cad", "power", "hr"),
  elevationShape = NA,
  elevationColor = "speed",
  showTime = TRUE,
  showSummary = TRUE,
  showStops = TRUE,
  stopToleranceMeters = 20,
  minSecsRolling = 10,
  minNumPoints = 3000,
  imperial = TRUE,
  ...
)
}
\arguments{
\item{track}{data frame or tibble containing track details including
location and elevation data, as well as heartrate and cadence}

\item{summary}{data frame or tibble containing track summary from the
data frames generated by \code{read_ride}}

\item{savefn}{name of output file, format is determined by the file
extension specified, .pdf, .png or .jpg}

\item{title}{title printed over top center of diagram}

\item{palette}{palette to use in graphic (\code{"plasma"} or \code{"magma"}
or \code{"inferno"} or \code{"viridis"})}

\item{naPlotColor}{name of color to use displaying NAs and zeros}

\item{verticalMultiplier}{default vertical exaggeration factor.  Default
ranges from 25 to 60 depending on the length being plotted}

\item{ppm}{override calculated default number of points per mile}

\item{displayBands}{character vector of variables displayed under profile,
in order from top to bottom}

\item{elevationShape}{shape to use for drawing the elevation plot}

\item{elevationColor}{variable to use to shade elevation line.  options are
"speed","power","hr","cad", or color name}

\item{showTime}{draw time axis - use FALSE to suppress,}

\item{showSummary}{display summary results if supplied}

\item{showStops}{draw marks for short and long stops on distance axis if
time axis not drawn}

\item{stopToleranceMeters}{tolerance in meters for determining stop}

\item{minSecsRolling}{restarts shorter than this number of seconds
considered to be part of stop}

\item{minNumPoints}{pad out the plot on the right if too short}

\item{imperial}{use mi and ft instead of km and m}

\item{...}{arguments for other functions including color bounds}
}
\value{
a ggplot object
}
\description{
\code{rideProfile} creates a plot which uses the elevation profile
  and colored bands to compactly display the relationship between terrain,
  speed, cadence and heartrate on a given ride
}
\details{
Generates a plot which compactly displays the elevation profile and its
   relationship with speed, cadence and heartrate.  Also displays clock
   information and shows location and approximate length of stops.
}
\section{main call }{

  \link{rideProfile}
}

