% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMisc.R
\name{setupColorBounds}
\alias{setupColorBounds}
\title{create functions to map variables into colors based on parameters}
\usage{
setupColorBounds(
  palette,
  legendwidth,
  speedLow = 3,
  speedHigh = 40,
  speedColorLow = 0,
  speedColorHigh = 40,
  gradeLow = -0.15,
  gradeHigh = 0.15,
  gradeColorLow = 40,
  gradeColorHigh = 0,
  hrLow = 100,
  hrHigh = 170,
  hrColorLow = 11,
  hrColorHigh = 26,
  cadCont = TRUE,
  cadTarget = 88,
  cadLow = 50,
  cadHigh = 120,
  cadColorLow = 4,
  cadColorMid = 10,
  cadColorHigh = 15,
  powerLow = 75,
  powerHigh = 400,
  powerColorLow = 9,
  powerColorHigh = 21,
  imperial = TRUE,
  ...
)
}
\arguments{
\item{palette}{viridis palette name, see main routine}

\item{legendwidth}{width of a legend chunk in distance units}

\item{speedLow}{speeds below this are same color}

\item{speedHigh}{speeds above this are same color}

\item{speedColorLow}{set color for speedLow and slower}

\item{speedColorHigh}{set color for speedHigh and faster}

\item{gradeLow}{grades below this are same color}

\item{gradeHigh}{grades above this are same color}

\item{gradeColorLow}{set color for gradeLow and steeper downhill}

\item{gradeColorHigh}{set color for gradeHigh and steeper uphill}

\item{hrLow}{heartrates below this are same color}

\item{hrHigh}{heartrates above this are same color}

\item{hrColorLow}{set color for hrLow and lower}

\item{hrColorHigh}{set color for hrHigh and higher
colors are from same palette as speeds, number is the speed corresponding
to the desired limit on the range of heartrates}

\item{cadCont}{display cadence as a continuos color map}

\item{cadTarget}{target cadence range minimum}

\item{cadLow}{lower cadence limit for continuous color, all
lower cadences are displayed as same color}

\item{cadHigh}{upper cadence limit for continuous color, all
higher cadences are displayed as same color}

\item{cadColorLow}{set color for cadence at cadLow or below}

\item{cadColorMid}{set color for cadence above low but below target}

\item{cadColorHigh}{set color for cadence above target}

\item{powerLow}{power outputs below this are same color}

\item{powerHigh}{power outputs above this are same color}

\item{powerColorLow}{set color for powerLow and lower}

\item{powerColorHigh}{set color for powerHigh and higher}

\item{imperial}{use mi and ft instead of km and m}

\item{...}{stuff for other functions}
}
\value{
a named list of named lists of functions
}
\description{
\code{setupColorBounds} creates lists of functions for color bars and legends
}
