% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMisc.R
\name{smoothSeries}
\alias{smoothSeries}
\title{create smoothed data series}
\usage{
smoothSeries(
  track,
  hrSmoothBW = 6,
  hrSmoothNN = 6,
  cadSmoothBW = 10,
  cadSmoothNN = 10,
  gradeSmoothBW = 13,
  gradeSmoothNN = 13,
  powerSmoothBW = 10,
  powerSmoothNN = 10,
  elevSmoothBWMeters = 15,
  imperial = TRUE,
  ...
)
}
\arguments{
\item{track}{data frame or tibble containing track details including
location and elevation data, as well as heartrate and cadence}

\item{hrSmoothBW}{bandwidth (in seconds) for smoothing kernel
for heartrate data}

\item{hrSmoothNN}{number of points (on each side) to use in smoothing
kernel for heartrate data}

\item{cadSmoothBW}{bandwidth (in seconds) for smoothing kernel
for cadence data}

\item{cadSmoothNN}{number of points (on each side) to use in
smoothing kernel for cadence data}

\item{gradeSmoothBW}{bandwidth (in seconds) for smoothing kernel
for grade data}

\item{gradeSmoothNN}{number of points (on each side) to use in
smoothing kernel for grade data}

\item{powerSmoothBW}{bandwidth (in seconds) for smoothing kernel
for power data}

\item{powerSmoothNN}{number of points (on each side) to use in smoothing
for power data}

\item{elevSmoothBWMeters}{bandwidth (in meters) for elevation smoothing}

\item{imperial}{use mi and ft instead of km and m}

\item{...}{arguments for other functions including color bounds}
}
\description{
\code{smoothSeries} creates lists of smoothed data series
}
