% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMain.R, R/rideProfile.R
\name{rideProfile}
\alias{rideProfile}
\title{plot heartrate, cadence, speed and more using color-driven
  data summaries to convey performance patterns}
\usage{
rideProfile(track, summary, savefn, title = "Ride starting ",
  palette = "plasma", naPlotColor = "gray88",
  verticalMultiplier = NA, ppm = NA, displayBands = c("speed",
  "grade", "cadence", "power", "heartrate"), elevationShape = 46,
  showTime = TRUE, showSummary = TRUE, showStops = TRUE,
  hrLow = 100, hrHigh = 170, hrColorLow = 11, hrColorHigh = 26,
  cadTarget = 88, cadCont = TRUE, cadLow = 50, cadHigh = 120,
  cadColorLow = 4, cadColorMid = 10, cadColorHigh = 15,
  powerLow = 75, powerHigh = 400, powerColorLow = 9,
  powerColorHigh = 21, hrSmoothBW = 6, hrSmoothNN = 6,
  cadSmoothBW = 10, cadSmoothNN = 10, gradeSmoothBW = 13,
  gradeSmoothNN = 13, powerSmoothBW = 10, powerSmoothNN = 10,
  elevSmoothBWMeters = 15, stopToleranceMeters = 20,
  minSecsRolling = 10, minNumPoints = 3000, imperial = TRUE, ...)
}
\arguments{
\item{track}{data frame or tibble containing track details including
location and elevation data, as well as heartrate and cadence}

\item{summary}{data frame or tibble containing track summary from the
data frames generated by \code{read_ride}}

\item{savefn}{name of output file, format is determined by the file
extension specified, .pdf, .png or .jpg}

\item{title}{title printed over top center of diagram}

\item{palette}{palette to use in graphic (\code{"plasma"} or \code{"magma"}
or \code{"inferno"} or \code{"viridis"})}

\item{naPlotColor}{name of color to use displaying NAs and zeros}

\item{verticalMultiplier}{default vertical exaggeration factor.  Default
ranges from 25 to 60 depending on the length being plotted}

\item{ppm}{override calculated default number of points per mile}

\item{displayBands}{character vector of variables displayed under profile,
in order from top to bottom}

\item{elevationShape}{shape to use for drawing the elevation plot}

\item{showTime}{draw time axis - use FALSE to suppress,}

\item{showSummary}{display summary results if supplied}

\item{showStops}{draw marks for short and long stops on distance axis if
time axis not drawn}

\item{hrLow}{heartrates below this are same color}

\item{hrHigh}{heartrates above this are same color}

\item{hrColorLow}{set color for hrLow and lower}

\item{hrColorHigh}{set color for hrHigh and higher
colors are from same palette as speeds, number is the speed corresponding
to the desired limit on the range of heartrates}

\item{cadTarget}{target cadence range minimum}

\item{cadCont}{display cadence as a continuos color map}

\item{cadLow}{lower cadence limit for continuous color, all
lower cadences are displayed as same color}

\item{cadHigh}{upper cadence limit for continuous color, all
higher cadences are displayed as same color}

\item{cadColorLow}{set color for cadence at cadLow or below}

\item{cadColorMid}{set color for cadence above low but below target}

\item{cadColorHigh}{set color for cadence above target}

\item{powerLow}{power outputs below this are same color}

\item{powerHigh}{power outputs above this are same color}

\item{powerColorLow}{set color for powerLow and lower}

\item{powerColorHigh}{set color for powerHigh and higher}

\item{hrSmoothBW}{bandwidth (in seconds) for smoothing kernel
for heartrate data}

\item{hrSmoothNN}{number of points (on each side) to use in smoothing
kernel for heartrate data}

\item{cadSmoothBW}{bandwidth (in seconds) for smoothing kernel
for cadence data}

\item{cadSmoothNN}{number of points (on each side) to use in
smoothing kernel for cadence data}

\item{gradeSmoothBW}{bandwidth (in seconds) for smoothing kernel
for grade data}

\item{gradeSmoothNN}{number of points (on each side) to use in
smoothing kernel for grade data}

\item{powerSmoothBW}{bandwidth (in seconds) for smoothing kernel
for power data}

\item{powerSmoothNN}{number of points (on each side) to use in smoothing
for power data}

\item{elevSmoothBWMeters}{bandwidth (in meters) for elevation smoothing}

\item{stopToleranceMeters}{tolerance in meters for determining stop}

\item{minSecsRolling}{restarts shorter than this number of seconds
considered to be part of stop}

\item{minNumPoints}{pad out the plot on the right if too short}

\item{imperial}{use mi and ft instead of km and m}

\item{...}{arguments for other functions}
}
\value{
a ggplot object
}
\description{
\code{rideProfile} creates a plot which uses the elevation profile
  and colored bands to compactly display the relationship between terrain,
  speed, cadence and heartrate on a given ride
}
\details{
Generates a plot which compactly displays the elevation profile and its
   relationship with speed, cadence and heartrate.  Also displays clock
   information and shows location and approximate length of stops.
}
\section{main call }{

  \link{rideProfile}
}

